linked list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stddef.h>
 #include <stdio.h>
 
 /**
@@ -9,7 +10,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int = count;
+	size_t = count;
 
 	count = 0;
 	while (h != NULL)
